+++ test/Control/Exception/AnnotatedSpec.hs
@@ -71,20 +71,44 @@ spec = do
                         pass
 
     describe "try" $ do
-        it "always adds a location" $ do
-            Left exn <- try (throw TestException)
-            exn `shouldBe` AnnotatedException [] TestException
+        let subject :: (Exception e, Exception e') => e -> IO e'
+            subject exn = do
+                Left exn' <- try (throw exn)
+                pure exn'
 
-        it "does not nest locations" $ do
-            Left exn <- try $ throw $ new $ new $ new TestException
-            exn `shouldBe` new TestException
+        describe "when throwing non-Annotated" $ do
+            it "can add an empty annotation for a non-Annotated exception" $ do
+                exn <- subject TestException
+                exn `shouldBe` AnnotatedException [] TestException
+
+            it "can catch a usual exception" $ do
+                exn <- subject TestException
+                exn `shouldBe` TestException
+
+        describe "when throwing Annotated" $ do
+            it "can catch a non-Annotated exception" $ do
+                exn <- subject $ new TestException
+                exn `shouldBe` TestException
+
+            it "can catch an Annotated exception" $ do
+                exn <- subject TestException
+                exn `shouldBe` new TestException
+
+        describe "nesting behavior" $ do
+            it "can catch at any level of nesting" $ do
+                subject TestException
+                    >>= (`shouldBe` new TestException)
+                subject TestException
+                    >>= (`shouldBe` new (new TestException))
+                subject TestException
+                    >>= (`shouldBe` new (new (new TestException)))
 
     describe "Safe.try" $ do
         it "can catch a located exception" $ do
             Left exn <- Safe.try (Safe.throw TestException)
             exn `shouldBe` new TestException
 
-        it "does not catch a AnnotatedException" $ do
+        it "does not catch an AnnotatedException" $ do
             let action = do
                     Left exn <- Safe.try (Safe.throw $ new TestException)
                     exn `shouldBe` TestException