+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 ## Unreleased changes
 
+## 0.1.1.0
+
+- [#4](https://github.com/parsonsmatt/annotated-exception/pull/4)
+    - Add `catches`
+    - Replace `Control.Exception.Safe.try` with `try` that can get an
+      `AnnotatedException e` or a regular, un-`Annotated` `e`.
+
 ## 0.1.0.0
 
 - Initial Release